[cmd] Deprecate --warnings flag
--warnings flag made "CodeChecker checkers" print clang-diagnostic...
checkers. Now this is deprecated, because all these warnings are
considered as normal checker names.

The corresponding codes have been moved under clang-tidy config
handler instead of the top-level file.
