+++ analyzer/codechecker_analyzer/cmd/checkers.py
@@ -13,7 +13,6 @@ List the checkers available in the analyzers.
 import argparse
 import json
 import os
-import subprocess
 import sys
 from collections import defaultdict
 from typing import Dict, Iterable, List, Tuple
@@ -22,8 +21,6 @@ from codechecker_report_converter import twodim
 
 from codechecker_analyzer import analyzer_context
 from codechecker_analyzer.analyzers import analyzer_types
-from codechecker_analyzer.analyzers.clangsa.analyzer import ClangSA
-from codechecker_analyzer.analyzers.clangtidy.analyzer import ClangTidy
 
 from codechecker_common import arg, logger
 from codechecker_common.output import USER_FORMATS
@@ -31,60 +28,6 @@ from codechecker_common.checker_labels import CheckerLabels
 from codechecker_analyzer.analyzers.config_handler import CheckerState
 
 LOG = logger.get_logger('system')
-
-
-def get_diagtool_bin():
-    """
-    Return full path of diagtool.
-
-    Select clang binary, check for a 'diagtool' binary next to the selected
-    clang binary and return full path of this binary if it exists.
-    """
-    context = analyzer_context.get_context()
-    clang_bin = context.analyzer_binaries.get(ClangSA.ANALYZER_NAME)
-
-    if not clang_bin:
-        return None
-
-    # Resolve symlink.
-    clang_bin = os.path.realpath(clang_bin)
-
-    # Find diagtool next to the clang binary.
-    diagtool_bin = os.path.join(os.path.dirname(clang_bin), 'diagtool')
-    if os.path.exists(diagtool_bin):
-        return diagtool_bin
-
-    LOG.debug("'diagtool' can not be found next to the clang binary (%s)!",
-              clang_bin)
-
-
-def get_warnings():
-    """
-    Returns list of warning flags by using diagtool.
-    """
-    diagtool_bin = get_diagtool_bin()
-
-    if not diagtool_bin:
-        return []
-
-    try:
-        result = subprocess.check_output(
-            [diagtool_bin, 'tree'],
-            env=analyzer_context.get_context().analyzer_env,
-            universal_newlines=True,
-            encoding="utf-8",
-            errors="ignore")
-        return [w[2:] for w in result.split() if w.startswith("-W")]
-    except subprocess.CalledProcessError as exc:
-        LOG.error("'diagtool' encountered an error while retrieving the "
-                  "checker list. If you are using a custom compiled clang, "
-                  "you may have forgotten to build the 'diagtool' target "
-                  "alongside 'clang' and 'clang-tidy'! Error message: %s",
-                  exc.output)
-
-        raise
-    except OSError:
-        raise
 
 
 def get_argparser_ctor_args():
@@ -159,13 +102,12 @@ def add_arguments_to_parser(parser):
                         help="Show checkers only from the analyzers "
                              "specified.")
 
-    if get_diagtool_bin():
-        parser.add_argument('-w', '--warnings',
-                            dest='show_warnings',
-                            default=argparse.SUPPRESS,
-                            action='store_true',
-                            required=False,
-                            help="Show available warning flags.")
+    parser.add_argument('-w', '--warnings',
+                        dest='show_warnings',
+                        default=argparse.SUPPRESS,
+                        action='store_true',
+                        required=False,
+                        help="DEPRECATED. Show available warning flags.")
 
     parser.add_argument('--details',
                         dest='details',
@@ -343,14 +285,6 @@ def __get_detailed_checker_info(
             checker_info[analyzer].append(
                 (state, checker, analyzer, description,
                  sorted(cl.labels_of_checker(checker, analyzer))))
-
-    if 'show_warnings' in args:
-        for warning in get_warnings():
-            warning = 'clang-diagnostic-' + warning
-            checker_info[ClangTidy.ANALYZER_NAME].append(
-                (CheckerState.default, warning, ClangTidy.ANALYZER_NAME, '',
-                 sorted(cl.labels_of_checker(
-                    warning, ClangTidy.ANALYZER_NAME))))
 
     return checker_info
