+++ analyzer/tests/functional/cmdline/test_cmdline.py
@@ -169,7 +169,17 @@ class TestCmdline(unittest.TestCase):
         self.assertEqual(retcode, 0)
         self.assertIn('clang-diagnostic-vla', out)
         # Make sure the header from `diagtool --tree` is ignored.
-        self.assertNotIn('EEN', out)
+        self.assertNotIn('GREEN', out)
+
+        # --warnings flag is deprecated. Warnings are included in checker list
+        # by default.
+        checkers_cmd = [env.codechecker_cmd(), 'checkers']
+        retcode, out, _ = run_cmd(checkers_cmd)
+
+        self.assertEqual(retcode, 0)
+        self.assertIn('clang-diagnostic-vla', out)
+        # Make sure the header from `diagtool --tree` is ignored.
+        self.assertNotIn('GREEN', out)
 
     def test_clangsa_checkers_description(self):
         """
